Refactor: Take out chat eaction handling out of main.py
- chat_logging.py: move handling reaction to responses here, since it doesn't really need to be in main.py for decluttering purposes
- constants.py: encapsulate into a class so that the variables will be shared and updated accordingly across modules not just in main
-main.py: Remove logging reaction handling here and use config class for USERNAME and ACCESS_TOKEN so it will always get the updated values when it is changed.
